Converted everything to AWT

+++ src/engine/EventProcessor.java
@@ -3,7 +3,7 @@ package engine;
 import java.util.Set;
 
 public class EventProcessor {
-	public void keysPressed(Set<Short> keys) {
+	public void keysPressed(Set<Integer> keys) {
 	}
 	
 	public void mouseMoved(int dx, int dy) {
